refact: missing library

+++ database/migrations/Subscription/2024_09_10_122123_add_payment_method_to_subscriptions_table.php
@@ -11,8 +11,15 @@ return new class () extends Migration {
     public function up(): void
     {
         Schema::table('subscriptions', function (Blueprint $table) {
-            $table->string('payment_method_id')->nullable()->after('apps_plans_id');
-            ;
+            // Check if 'apps_plans_id' column doesn't exist, then add it as a big integer
+            if (! Schema::hasColumn('subscriptions', 'apps_plans_id')) {
+                $table->unsignedBigInteger('apps_plans_id')->nullable();
+            }
+
+            // Check if 'payment_method_id' column doesn't exist, then add it
+            if (! Schema::hasColumn('subscriptions', 'payment_method_id')) {
+                $table->string('payment_method_id')->nullable()->after('apps_plans_id');
+            }
         });
     }
 
@@ -22,7 +29,15 @@ return new class () extends Migration {
     public function down(): void
     {
         Schema::table('subscriptions', function (Blueprint $table) {
-            $table->dropColumn('payment_method_id');
+            // Drop 'payment_method_id' if it exists
+            if (Schema::hasColumn('subscriptions', 'payment_method_id')) {
+                $table->dropColumn('payment_method_id');
+            }
+
+            // Drop 'apps_plans_id' if it exists
+            if (Schema::hasColumn('subscriptions', 'apps_plans_id')) {
+                $table->dropColumn('apps_plans_id');
+            }
         });
     }
 };
